Split netcdf_engine/converter.py into multiple files
Refactor to improve readability of code base.
