+++ tiledb/cf/netcdf_engine/api.py
@@ -1,5 +1,7 @@
 # Copyright 2021 TileDB Inc.
 # Licensed under the MIT License.
+"""Functions for converting NetCDF files to TileDB."""
+
 from pathlib import Path
 from typing import Dict, Optional, Sequence, Union
