+++ tiledb/cf/netcdf_engine/converter.py
@@ -2,583 +2,32 @@
 # Licensed under the MIT License.
 """Classes for converting NetCDF4 files to TileDB."""
 
-import time
 import warnings
-from abc import abstractmethod
 from collections import defaultdict
-from contextlib import contextmanager
 from io import StringIO
 from pathlib import Path
-from typing import Any, Dict, List, Optional, Sequence, Tuple, Union
+from typing import Dict, List, Optional, Sequence, Tuple, Union
 
 import netCDF4
 import numpy as np
 
 import tiledb
-from tiledb.cf.core import AttrMetadata, DimMetadata, Group
-from tiledb.cf.creator import (
-    ArrayCreator,
-    ArrayRegistry,
-    AttrCreator,
-    DataspaceCreator,
-    DataspaceRegistry,
-    DType,
-    SharedDim,
+from tiledb.cf.core import Group
+from tiledb.cf.creator import ArrayCreator, DataspaceCreator, DataspaceRegistry
+
+from ._attr_converters import NetCDF4ToAttrConverter, NetCDF4VarToAttrConverter
+from ._dim_converters import (
+    NetCDF4CoordToDimConverter,
+    NetCDF4DimToDimConverter,
+    NetCDF4ScalarToDimConverter,
+    NetCDF4ToDimConverter,
 )
-
-_DEFAULT_INDEX_DTYPE = np.dtype("uint64")
-COORDINATE_SUFFIX = ".data"
-
-
-class NetCDF4ToAttrConverter(AttrCreator):
-    """Abstract base class for classes that copy data from objects in a NetCDF group to
-    a TileDB attribute.
-    """
-
-    def copy_metadata(self, netcdf_group: netCDF4.Dataset, tiledb_array: tiledb.Array):
-        """Copy the metadata data from NetCDF to TileDB.
-
-        Parameters:
-            netcdf_group: NetCDF group to get the metadata items from.
-            tiledb_array: TileDB array to copy the metadata items to.
-        """
-
-    @abstractmethod
-    def get_values(
-        self, netcdf_group: netCDF4.Dataset, sparse: bool
-    ) -> Union[np.ndarray, slice]:
-        """Returns values from a NetCDF group that will be copied to TileDB.
-
-        Parameters:
-            netcdf_group: NetCDF group to get the values from.
-            sparse: ``True`` if copying into a sparse array and ``False`` if copying
-                into a dense array.
-
-        Returns:
-            The coordinates needed for querying the create TileDB dimension in the form
-                of a numpy array if sparse is ``True`` and a slice otherwise.
-        """
-
-
-class NetCDF4ToDimConverter(SharedDim):
-    """Abstract base class for classes that copy data from objects in a NetCDF group to
-    a TileDB dimension.
-    """
-
-    def copy_metadata(self, netcdf_group: netCDF4.Dataset, tiledb_array: tiledb.Array):
-        """Copy the metadata data from NetCDF to TileDB.
-
-        Parameters:
-            netcdf_group: NetCDF group to get the metadata items from.
-            tiledb_array: TileDB array to copy the metadata items to.
-        """
-
-    @abstractmethod
-    def get_values(
-        self, netcdf_group: netCDF4.Dataset, sparse: bool
-    ) -> Union[np.ndarray, slice]:
-        """Returns values from a NetCDF group that will be copied to TileDB.
-
-        Parameters:
-            netcdf_group: NetCDF group to get the values from.
-            sparse: ``True`` if copying into a sparse array and ``False`` if copying
-                into a dense array.
-
-        Returns:
-            The coordinates needed for querying the create TileDB dimension in the form
-                of a numpy array if sparse is ``True`` and a slice otherwise.
-        """
-
-
-class NetCDF4CoordToDimConverter(NetCDF4ToDimConverter):
-    """Converter for a NetCDF variable/dimension pair to a TileDB dimension.
-
-    Attributes:
-        name: Name of the TileDB dimension.
-        domain: The (inclusive) interval on which the dimension is valid.
-        dtype: The numpy dtype of the values and domain of the dimension.
-        input_dim_name: The name of input NetCDF dimension.
-        input_var_name: The name of input NetCDF variable.
-        input_var_dtype: The numpy dtype of the input NetCDF variable.
-    """
-
-    def __init__(
-        self,
-        dataspace_registry: DataspaceRegistry,
-        name: str,
-        domain: Optional[Tuple[Optional[DType], Optional[DType]]],
-        dtype: np.dtype,
-        input_dim_name: str,
-        input_var_name: str,
-        input_var_dtype: np.dtype,
-    ):
-        super().__init__(dataspace_registry, name, domain, dtype)
-        self.input_dim_name = input_dim_name
-        self.input_var_name = input_var_name
-        self.input_var_dtype = input_var_dtype
-
-    def __eq__(self, other):
-        return (
-            super.__eq__(self, other)
-            and self.input_dim_name == other.input_dim_name
-            and self.input_var_name == other.input_var_name
-            and self.input_var_dtype == other.input_var_dtype
-        )
-
-    def __repr__(self):
-        return (
-            f"NetCDFVariable(name={self.input_var_name}, dtype={self.input_var_dtype}) "
-            f" -> {super().__repr__()}"
-        )
-
-    def html_input_summary(self):
-        """Returns a HTML string summarizing the input for the dimension."""
-        return (
-            f"NetCDFVariable(name={self.input_var_name}, dtype={self.input_var_dtype})"
-        )
-
-    @classmethod
-    def from_netcdf(
-        cls,
-        dataspace_registry: DataspaceRegistry,
-        var: netCDF4.Variable,
-        name: Optional[str] = None,
-        domain: Optional[Tuple[DType, DType]] = None,
-    ):
-        if len(var.dimensions) != 1:
-            raise ValueError(
-                f"Cannot create dimension from variable '{var.name}' with shape "
-                f"{var.shape}. Coordinate variables must have only one dimension."
-            )
-        add_offset = get_ncattr(var, "add_offset")
-        scale_factor = get_ncattr(var, "scale_factor")
-        unsigned = get_ncattr(var, "_Unsigned")
-        if add_offset is not None or scale_factor is not None or unsigned is not None:
-            raise NotImplementedError(
-                f"Cannot convert variable {var.name} into a TileDB dimension. Support "
-                f"for converting scaled coordinates has not yet been implemented."
-            )
-        dtype = np.dtype(var.dtype)
-        return cls(
-            dataspace_registry=dataspace_registry,
-            name=name if name is not None else var.name,
-            domain=domain,
-            dtype=dtype,
-            input_dim_name=var.dimensions[0],
-            input_var_name=var.name,
-            input_var_dtype=dtype,
-        )
-
-    def copy_metadata(self, netcdf_group: netCDF4.Dataset, tiledb_array: tiledb.Array):
-        """Copy the metadata data from NetCDF to TileDB.
-
-        Parameters:
-            netcdf_group: NetCDF group to get the metadata items from.
-            tiledb_array: TileDB array to copy the metadata items to.
-        """
-        try:
-            variable = netcdf_group.variables[self.input_var_name]
-        except KeyError as err:
-            raise KeyError(
-                f"The variable '{self.input_var_name}' was not found in the provided "
-                f"NetCDF group."
-            ) from err
-        dim_meta = DimMetadata(tiledb_array.meta, self.name)
-        for key in variable.ncattrs():
-            safe_set_metadata(dim_meta, key, variable.getncattr(key))
-
-    def get_values(
-        self,
-        netcdf_group: netCDF4.Dataset,
-        sparse: bool,
-    ):
-        """Returns the values of the NetCDF coordinate that is being copied, or
-        None if the coordinate is of size 0.
-
-        Parameters:
-            netcdf_group: NetCDF group to get the coordinate values from.
-            sparse: ``True`` if copying into a sparse array and ``False`` if copying
-                into a dense array.
-
-        Returns:
-            The coordinate values needed for querying the TileDB dimension in the
-                form a numpy array.
-        """
-        if not sparse:
-            raise NotImplementedError(
-                "Support for copying NetCDF coordinates to dense arrays has not "
-                "been implemented."
-            )
-        try:
-            variable = netcdf_group.variables[self.input_var_name]
-        except KeyError as err:
-            raise KeyError(
-                f"The variable '{self.input_var_name}' was not found in the provided "
-                f"NetCDF group."
-            ) from err
-        if variable.ndim != 1:
-            raise ValueError(
-                f"The variable '{self.input_var_name}' with {variable.ndim} dimensions "
-                f"is not a valid NetCDF coordinate. Cannot copy data from variable "
-                f"'{self.input_var_name}' to TileDB dimension '{self.name}'."
-            )
-        if variable.get_dims()[0].size < 1:
-            return None
-        return variable[:]
-
-    @property
-    def input_name(self) -> str:
-        """(DEPRECATED) Name of the input NetCDF variable and dimension."""
-        with warnings.catch_warnings():
-            warnings.warn(
-                "Deprecated. Use `input_var_name` to get the name of the input NetCDF "
-                "variable and `input_dim_name` to get the input NetCDF dimension.",
-                DeprecationWarning,
-            )
-        if self.input_var_name != self.input_dim_name:
-            raise ValueError(
-                "Input name is ambiguous. The input variable and input dimension have "
-                "different names."
-            )
-        return self.input_var_name
-
-    @property
-    def input_dtype(self) -> np.dtype:
-        """(DEPRECATED) Name of the input NetCDF variable and dimension."""
-        with warnings.catch_warnings():
-            warnings.warn(
-                "Deprecated. Use `input_var_dtype` instead.",
-                DeprecationWarning,
-            )
-        return self.input_var_dtype
-
-    @property
-    def is_index_dim(self) -> bool:
-        return False
-
-
-class NetCDF4DimToDimConverter(NetCDF4ToDimConverter):
-    """Converter for a NetCDF dimension to a TileDB dimension.
-
-    Attributes:
-        name: Name of the TileDB dimension.
-        domain: The (inclusive) interval on which the dimension is valid.
-        dtype: The numpy dtype of the values and domain of the dimension.
-        input_dim_name: Name of the input NetCDF variable.
-        input_dim_size: Size of the input NetCDF variable.
-        is_unlimited: If True, the input NetCDF variable is unlimited.
-    """
-
-    def __init__(
-        self,
-        dataspace_registry: DataspaceRegistry,
-        name: str,
-        domain: Optional[Tuple[Optional[DType], Optional[DType]]],
-        dtype: np.dtype,
-        input_dim_name: str,
-        input_dim_size: int,
-        is_unlimited: bool,
-    ):
-        super().__init__(dataspace_registry, name, domain, dtype)
-        self.input_dim_name = input_dim_name
-        self.input_dim_size = input_dim_size
-        self.is_unlimited = is_unlimited
-
-    def __eq__(self, other):
-        return (
-            super.__eq__(self, other)
-            and self.input_dim_name == other.input_dim_name
-            and self.input_dim_size == other.input_dim_size
-            and self.is_unlimited == other.is_unlimited
-        )
-
-    def __repr__(self):
-        size_str = "unlimited" if self.is_unlimited else str(self.input_dim_size)
-        return (
-            f"NetCDFDimension(name={self.input_dim_name}, size={size_str}) -> "
-            f"{super().__repr__()}"
-        )
-
-    def html_input_summary(self):
-        """Returns a HTML string summarizing the input for the dimension."""
-        size_str = "unlimited" if self.is_unlimited else str(self.input_dim_size)
-        return f"NetCDFDimension(name={self.input_dim_name}, size={size_str})"
-
-    @classmethod
-    def from_netcdf(
-        cls,
-        dataspace_registry: DataspaceRegistry,
-        dim: netCDF4.Dimension,
-        unlimited_dim_size: Optional[int],
-        dtype: np.dtype,
-        name: Optional[str] = None,
-    ):
-        size = (
-            unlimited_dim_size
-            if dim.isunlimited() and unlimited_dim_size is not None
-            else dim.size
-        )
-        return cls(
-            dataspace_registry=dataspace_registry,
-            name=name if name is not None else dim.name,
-            domain=(0, size - 1),
-            dtype=dtype,
-            input_dim_name=dim.name,
-            input_dim_size=dim.size,
-            is_unlimited=dim.isunlimited(),
-        )
-
-    def get_values(
-        self, netcdf_group: netCDF4.Dataset, sparse: bool
-    ) -> Union[np.ndarray, slice]:
-        """Returns the values of the NetCDF dimension that is being copied.
-
-        Parameters:
-            netcdf_group: NetCDF group to get the dimension values from.
-            sparse: ``True`` if copying into a sparse array and ``False`` if copying
-                into a dense array.
-
-        Returns:
-            The coordinates needed for querying the created TileDB dimension in the form
-                of a numpy array if sparse is ``True`` and a slice otherwise.
-        """
-        group = netcdf_group
-        while group is not None:
-            if self.input_dim_name in group.dimensions:
-                dim = group.dimensions[self.input_dim_name]
-                if dim.size == 0:
-                    raise ValueError(
-                        f"Cannot copy dimension data from NetCDF dimension "
-                        f"'{self.input_dim_name}' to TileDB dimension '{self.name}'. "
-                        f"The NetCDF dimension is of size 0; there is no data to copy."
-                    )
-                if self.domain is not None and (
-                    self.domain[1] is not None and dim.size - 1 > self.domain[1]
-                ):
-                    raise IndexError(
-                        f"Cannot copy dimension data from NetCDF dimension "
-                        f"'{self.input_dim_name}' to TileDB dimension '{self.name}'. "
-                        f"The NetCDF dimension size of {dim.size} does not fit in the "
-                        f"domain {self.domain} of the TileDB dimension."
-                    )
-                if sparse:
-                    return np.arange(dim.size)
-                return slice(dim.size)
-            group = group.parent
-        raise KeyError(
-            f"Unable to copy NetCDF dimension '{self.input_dim_name}' to the TileDB "
-            f"dimension '{self.name}'. No NetCDF dimension with that name exists in "
-            f"the NetCDF group '{netcdf_group.path}' or its parent groups."
-        )
-
-    @property
-    def input_name(self) -> str:
-        """(DEPRECATED) Name of the input NetCDF dimension."""
-        with warnings.catch_warnings():
-            warnings.warn(
-                "Deprecated. Use `input_dim_name` instead.",
-                DeprecationWarning,
-            )
-        return self.input_dim_name
-
-    @property
-    def input_size(self) -> int:
-        """(DEPRECATED) Size of the input NetCDF dimension."""
-        with warnings.catch_warnings():
-            warnings.warn(
-                "Deprecated. Use `input_dim_size` instead.",
-                DeprecationWarning,
-            )
-        return self.input_dim_size
-
-
-class NetCDF4ScalarToDimConverter(NetCDF4ToDimConverter):
-    """Converter for NetCDF scalar (empty) dimensions to a TileDB Dimension.
-
-    Attributes:
-        name: Name of the TileDB dimension.
-        domain: The (inclusive) interval on which the dimension is valid.
-        dtype: The numpy dtype of the values and domain of the dimension.
-    """
-
-    def __repr__(self):
-        return f" Scalar dimensions -> {super().__repr__()}"
-
-    def html_input_summary(self):
-        """Returns a string HTML summary."""
-        return "NetCDF empty dimension"
-
-    def get_values(
-        self, netcdf_group: netCDF4.Dataset, sparse: bool
-    ) -> Union[np.ndarray, slice]:
-        """Get dimension values from a NetCDF group.
-
-        Parameters:
-            netcdf_group: NetCDF group to get the dimension values from.
-            sparse: ``True`` if copying into a sparse array and ``False`` if copying
-                into a dense array.
-
-        Returns:
-            The coordinates needed for querying the create TileDB dimension in the form
-                of a numpy array if sparse is ``True`` and a slice otherwise.
-        """
-        if sparse:
-            return np.array([0])
-        return slice(1)
-
-    @classmethod
-    def create(
-        cls, dataspace_registry: DataspaceRegistry, dim_name: str, dtype: np.dtype
-    ):
-        return cls(dataspace_registry, dim_name, (0, 0), dtype)
-
-
-class NetCDF4VarToAttrConverter(NetCDF4ToAttrConverter):
-    """Converter for a NetCDF variable to a TileDB attribute.
-
-    Attributes:
-        name: Name of the new attribute.
-        dtype: Numpy dtype of the attribute.
-        fill: Fill value for unset cells.
-        var: Specifies if the attribute is variable length (automatic for
-            byte/strings).
-        nullable: Specifies if the attribute is nullable using validity tiles.
-        filters: Specifies compression filters for the attribute.
-        input_var_name: Name of the input NetCDF variable that will be converted.
-        input_var_dtype: Numpy dtype of the input NetCDF variable.
-    """
-
-    def __init__(
-        self,
-        array_registry: ArrayRegistry,
-        name: str,
-        dtype: np.dtype,
-        fill: Optional[Union[int, float, str]],
-        var: bool,
-        nullable: bool,
-        filters: Optional[tiledb.FilterList],
-        input_var_name: str,
-        input_var_dtype: np.dtype,
-    ):
-        super().__init__(
-            array_registry=array_registry,
-            name=name,
-            dtype=dtype,
-            fill=fill,
-            var=var,
-            nullable=nullable,
-            filters=filters,
-        )
-        self.input_var_name = input_var_name
-        self.input_var_dtype = input_var_dtype
-
-    def __repr__(self):
-        return (
-            f"NetCDFVariable(name={self.input_var_name}, dtype={self.input_var_dtype})"
-            f" -> {super().__repr__()}"
-        )
-
-    def copy_metadata(self, netcdf_group: netCDF4.Dataset, tiledb_array: tiledb.Array):
-        """Copy the metadata data from NetCDF to TileDB.
-
-        Parameters:
-            netcdf_group: NetCDF group to get the metadata items from.
-            tiledb_array: TileDB array to copy the metadata items to.
-        """
-        try:
-            variable = netcdf_group.variables[self.input_var_name]
-        except KeyError as err:
-            raise KeyError(
-                f"The variable '{self.input_var_name}' was not found in the provided "
-                f"NetCDF group."
-            ) from err
-        attr_meta = AttrMetadata(tiledb_array.meta, self.name)
-        for key in variable.ncattrs():
-            safe_set_metadata(attr_meta, key, variable.getncattr(key))
-
-    def html_summary(self):
-        """Returns a string HTML summary."""
-        return (
-            f"NetCDFVariable(name={self.input_var_name}, dtype={self.input_var_dtype})"
-            f"{super().html_summary()}"
-        )
-
-    @classmethod
-    def from_netcdf(
-        cls,
-        array_registry: ArrayRegistry,
-        ncvar: netCDF4.Variable,
-        name: Optional[str] = None,
-        dtype: Optional[np.dtype] = None,
-        fill: Optional[Union[int, float, str]] = None,
-        var: bool = False,
-        nullable: bool = False,
-        filters: Optional[tiledb.FilterList] = None,
-    ):
-        if fill is None and "_FillValue" in ncvar.ncattrs():
-            fill = ncvar.getncattr("_FillValue")
-        if name is None:
-            name = (
-                ncvar.name
-                if ncvar.name not in ncvar.dimensions
-                else ncvar.name + COORDINATE_SUFFIX
-            )
-        dtype = np.dtype(dtype) if dtype is not None else np.dtype(ncvar.dtype)
-        return cls(
-            array_registry=array_registry,
-            name=name,
-            dtype=dtype,
-            fill=fill,
-            var=var,
-            nullable=nullable,
-            filters=filters,
-            input_var_name=ncvar.name,
-            input_var_dtype=ncvar.dtype,
-        )
-
-    def get_values(
-        self, netcdf_group: netCDF4.Dataset, sparse: bool
-    ) -> Union[np.ndarray, slice]:
-        """Returns TileDB attribute values from a NetCDF group.
-
-        Parameters:
-            netcdf_group: NetCDF group to get the dimension values from.
-            sparse: ``True`` if copying into a sparse array and ``False`` if copying
-                into a dense array.
-
-        Returns:
-            The values needed to set an attribute in a TileDB array. If the array
-        is sparse the values will be returned as an 1D array; otherwise, they will
-        be returned as an ND array.
-        """
-        try:
-            variable = netcdf_group.variables[self.input_var_name]
-        except KeyError as err:
-            raise KeyError(
-                f"The variable '{self.input_var_name}' was not found in the provided "
-                f"NetCDF group."
-            ) from err
-        return variable[...].flatten() if sparse else variable[...]
-
-    @property
-    def input_name(self) -> str:
-        """(DEPRECATED) Name of the input NetCDF variable."""
-        with warnings.catch_warnings():
-            warnings.warn(
-                "Deprecated. Use `input_var_name` instead.",
-                DeprecationWarning,
-            )
-        return self.input_var_name
-
-    @property
-    def input_dtype(self) -> np.dtype:
-        """(DEPRECATED) Size of the input NetCDF dimension."""
-        with warnings.catch_warnings():
-            warnings.warn(
-                "Deprecated. Use `input_var_dtype` instead.",
-                DeprecationWarning,
-            )
-        return self.input_var_dtype
+from ._utils import (
+    _DEFAULT_INDEX_DTYPE,
+    copy_group_metadata,
+    get_variable_chunks,
+    open_netcdf_group,
+)
 
 
 class NetCDF4ArrayConverter(ArrayCreator):
@@ -1489,82 +938,3 @@ class NetCDF4ConverterEngine(DataspaceCreator):
                 if isinstance(array_creator, NetCDF4ArrayConverter):
                     with tiledb.open(array_uri, mode="w", key=key, ctx=ctx) as array:
                         array_creator.copy(netcdf_group, array)
-
-
-def copy_group_metadata(netcdf_group: netCDF4.Group, meta: tiledb.libtiledb.Metadata):
-    """Copy all NetCDF group attributs to a the metadata in a TileDB array."""
-    for key in netcdf_group.ncattrs():
-        value = netcdf_group.getncattr(key)
-        if key == "history":
-            value = f"{value} - TileDB array created on {time.ctime(time.time())}"
-        safe_set_metadata(meta, key, value)
-
-
-def safe_set_metadata(meta, key, value):
-    """Copy a metadata item to a TileDB array catching any errors as warnings."""
-    if isinstance(value, np.ndarray):
-        value = tuple(value.tolist())
-    elif isinstance(value, np.generic):
-        value = (value.tolist(),)
-    try:
-        meta[key] = value
-    except ValueError as err:  # pragma: no cover
-        with warnings.catch_warnings():
-            warnings.warn(f"Failed to set metadata `{key}={value}` with error: {err}")
-
-
-def get_ncattr(netcdf_item, key: str) -> Any:
-    if key in netcdf_item.ncattrs():
-        return netcdf_item.getncattr(key)
-    return None
-
-
-def get_variable_chunks(variable: netCDF4.Variable) -> Optional[Tuple[int, ...]]:
-    chunks = variable.chunking()
-    return None if chunks is None or chunks == "contiguous" else tuple(chunks)
-
-
-@contextmanager
-def open_netcdf_group(
-    group: Optional[Union[netCDF4.Dataset, netCDF4.Group]] = None,
-    input_file: Optional[Union[str, Path]] = None,
-    group_path: Optional[str] = None,
-):
-    """Context manager for opening a NetCDF group.
-
-    If both an input file and group are provided, this function will prioritize
-    opening from the group.
-
-    Parameters:
-        group: A NetCDF group to read from.
-        input_file: A NetCDF file to read from.
-        group_path: The path to the NetCDF group to read from in a NetCDF file. Use
-            ``'/'`` to specify the root group.
-    """
-    if group is not None:
-        if not isinstance(group, (netCDF4.Dataset, netCDF4.Group)):
-            raise TypeError(
-                f"Invalid input: group={group} of type {type(group)} is not a netCDF "
-                f"Group or or Dataset."
-            )
-        yield group
-    else:
-        if input_file is None:
-            raise ValueError(
-                "An input file must be provided; no default input file was set."
-            )
-        if group_path is None:
-            raise ValueError(
-                "A group path must be provided; no default group path was set. Use "
-                "``'/'`` for the root group."
-            )
-        root_group = netCDF4.Dataset(input_file)
-        root_group.set_auto_maskandscale(False)
-        try:
-            netcdf_group = root_group
-            if group_path != "/":
-                for child_group_name in group_path.strip("/").split("/"):
-                    netcdf_group = netcdf_group.groups[child_group_name]
-            yield netcdf_group
-        finally:
-            root_group.close()
